+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### v1.0.3(May 13, 2019)
+* Added `reopenWithCompletionHandler:` in `SBDSKTicket` class.
+* Added `createTicketWithTitle:userName:groupKey:customField:completionHandler:` in `SBDSKTicket` class.
+* Added `getOpenedListWithOffset:customFieldFilter:completionHandler:` in `SBDSKTicket` class.
+* Added `getClosedListWithOffset:customFieldFilter:completionHandler:` in `SBDSKTicket` class.
+
 ### v1.0.2(Mar 30, 2018)
 * Support SendBird Desk standalone.